feat: add new environments `eq*` and `eq`
This commit defines two new environments `eq*` and `eq` to "re-define"
the original `equation*` and `equation` - adjusting their top margins so
that they have the same height of space above and below.

Note that this change doesn't affect the original environments, meaning
that you can still use the premitive ones, or use the display math
delimiters "\[", "\]".

And we attempted to introduce new commands `\eq*` and `\eq` to simplify
the usage of the environments, but the XeLaTeX compiler panicked and
returned "You can't use `macro parameter character #' in math mode". We
are still trying to make them work.